requirements.txt
asgiref==3.5.2
Django==4.0.5
django-cors-headers==3.13.0
python-dotenv==0.20.0
sqlparse==0.4.2
